fix change distance in addresses module

+++ src/sql-queries/street.js
@@ -66,7 +66,7 @@ const CHANGE_STREET = `
 		),
 		street_distance = (
 			CASE
-				WHEN $3 > 0 THEN $3
+				WHEN $3::NUMERIC > 0 THEN ($3)::NUMERIC
 				ELSE s.street_distance
 			END
 		)
